requirements.txt
Flask
Flask-Mail
psycopg2-binary
APScheduler
requests
python-dotenv
flask-cors
Flask_Limiter